Fyritius, Blood Collection, Mentions of Alch Shop
Fyritius Brew really needs a sprite. Would seriously love help with that. In regards to everything else however, players can collect blood as long as a hand is revealed and has a cut on the limb. CUTS SPECIFICALLY, Slashing Wounds, you cannot stab yourself and collect blood that way, puncture wounds do not bleed as much as a wide cut.

This is to lead into the eventual addition of "Blood-Infused Health Pots", which improves the quality of the potion at the cost of one's own blood. This is to help prevent spamming a "greater potion" and will lead to asking for donations to create stronger health pots (or yknow just drink water which I really wanna change.)