The Great Big Religion Revert/Update Part 1
So! The Church is in a bad place in more then one way right now - Their miracles are weak and are very easily replicated by potions, surgery or even arcane spell casting, it's lore is being taken in a direction that doesn't align with the server vision and outside of some spiriting really haven't gotten the love they deserve. So lets fix that!

Changes include:
- Reverts Miracle to Lesser Miracle, it's original Ratwood version. Healing is a flat 25 (no more giant messy if trees) and it comes out instantly and onto a single body part.
- Fortify returns to Miracle
- Both lesser miracle and miracle have their fire damage and stun back. This may need to be tuned!
- The return of churn, huzzah!
- Removed abrogation from Necra's spell list.
- Removed the chance of creating a rotman via Remove Rot use.