Templar Weapon Fix
There's currently a block of code in the templar role that is supposed to give them a choice of weapon and a slight boost to that skill, putting them on par with the average melee class. This did not work. These errors have been corrected.

This bug existed in Azure since early September - it's still not fixed yet :^)