Removes the Female Exclusive Power of Bikini Armor (#100)
* Update armor.dm

Added males to the list of allowed sexes to wear bikini armor, and added the mob overlay sprites directory to each of the bikini armor pieces so they have their male counterparts to point to.

* Create armor.dm

Adds armor.dm file containing code for the egyptian robe under the name 'foreign dress' as well as code for barrel armor (only admin spawnable for now)

* Create weaving.dm

Adds weaving.dm for crafting the egyptian robes and headdresses

* Create armor.dmi

Placeholder armor.dmi that will hold the inventory icon for the barrel armor

* Create armor.dmi

Placeholder armor.dmi that will hold the onmob sprites for the new barrel armor, egyptian robe, and male bikini sprites.

* Add files via upload

Uploaded the barrel armor inventory icon .dmi and the onmob sprites .dmi

* Update armor.dm

Fixed Indentation error that I'm not even sure how it occurred when I copied and pasted the code.

* Reuploading file

Reuploading file because it didn't look like it took it the first time I tried to upload it for whatever reason.

* I SAID UPLOAD