Test Merge - Quiver fix (#107)
* Test Merge - Quiver fix

This should fix an issue with the quivers causing StrongDMM to be unstable. Tested locally by two separate testers. Please test merge this!

* Update traits.dm

* Merges be merging

* Update musketeer.dm

* Fix n' Nerf

Corrected the ammo pouch going into the wrong slot, preventing it from spawning. Nerfed the ranged skills of the boltman as, with their weapon choice, they would have 5 (!) skill in their weapon of choice.

* Update townguard.dm

And now the fix for the town guard

* Fine keep your gorgets